--- a/lib/core/base/virtual-node/virtual-node.js
+++ b/lib/core/base/virtual-node/virtual-node.js
@@ -55,24 +55,28 @@ class VirtualNode extends AbstractVirtualNode {
   // abstract Node properties so we can run axe in DOM-less environments.
   // add to the prototype so memory is shared across all virtual nodes
   get props() {
-    const {
-      nodeType,
-      nodeName,
-      id,
-      multiple,
-      nodeValue,
-      value
-    } = this.actualNode;
+    if (!this._cache.hasOwnProperty('props')) {
+      const {
+        nodeType,
+        nodeName,
+        id,
+        multiple,
+        nodeValue,
+        value
+      } = this.actualNode;
 
-    return {
-      nodeType,
-      nodeName: this._isXHTML ? nodeName : nodeName.toLowerCase(),
-      id,
-      type: this._type,
-      multiple,
-      nodeValue,
-      value
-    };
+      this._cache.props = {
+        nodeType,
+        nodeName: this._isXHTML ? nodeName : nodeName.toLowerCase(),
+        id,
+        type: this._type,
+        multiple,
+        nodeValue,
+        value
+      };
+    }
+
+    return this._cache.props;
   }
 
   /**