--- a/lib/core/utils/matches.js
+++ b/lib/core/utils/matches.js
@@ -211,6 +211,50 @@ export function convertSelector(selector) {
 }
 
 /**
+ * Determine if a virtual node matches a Slick format CSS expression.
+ *
+ * Note: this function is in the hot path, avoid memory allocation.
+ *
+ * @private
+ * @method optimizedMatchesExpression
+ * @memberof axe.utils
+ * @param {VirtualNode} vNode VirtualNode to match
+ * @param {Object|Object[]} expressions CSS selector expression or array of expressions
+ * @param {Number|NaN} index index of expression to match on if expressions is array
+ * @returns {Boolean}
+ */
+function optimizedMatchesExpression(vNode, expressions, index, matchAnyParent) {
+  const isArray = Array.isArray(expressions);
+  const expression = isArray ? expressions[index] : expressions;
+  let matches = matchExpression(vNode, expression);
+
+  while (!matches && matchAnyParent && vNode.parent) {
+    vNode = vNode.parent;
+    matches = matchExpression(vNode, expression);
+  }
+
+  if (index > 0) {
+    if ([' ', '>'].includes(expression.combinator) === false) {
+      throw new Error(
+        'axe.utils.matchesExpression does not support the combinator: ' +
+          expression.combinator
+      );
+    }
+
+    matches =
+      matches &&
+      optimizedMatchesExpression(
+        vNode.parent,
+        expressions,
+        index - 1,
+        expression.combinator === ' '
+      );
+  }
+
+  return matches;
+}
+
+/**
  * Determine if a virtual node matches a Slick format CSS expression
  *
  * @private
@@ -221,29 +265,12 @@ export function convertSelector(selector) {
  * @returns {Boolean}
  */
 export function matchesExpression(vNode, expressions, matchAnyParent) {
-  const exps = [].concat(expressions);
-  const expression = exps.pop();
-  let matches = matchExpression(vNode, expression);
-
-  while (!matches && matchAnyParent && vNode.parent) {
-    vNode = vNode.parent;
-    matches = matchExpression(vNode, expression);
-  }
-
-  if (exps.length) {
-    if ([' ', '>'].includes(expression.combinator) === false) {
-      throw new Error(
-        'axe.utils.matchesExpression does not support the combinator: ' +
-          expression.combinator
-      );
-    }
-
-    matches =
-      matches &&
-      matchesExpression(vNode.parent, exps, expression.combinator === ' ');
-  }
-
-  return matches;
+  return optimizedMatchesExpression(
+    vNode,
+    expressions,
+    expressions.length - 1,
+    matchAnyParent
+  );
 }
 
 /**