--- a/lib/core/utils/query-selector-all-filter.js
+++ b/lib/core/utils/query-selector-all-filter.js
@@ -26,13 +26,15 @@ function matchExpressions(domTree, expressions, filter) {
     const vNode = currentLevel.vNodes.pop();
     const childOnly = []; // we will add hierarchical '>' selectors here
     const childAny = [];
-    const combined = currentLevel.anyLevel
-      .slice()
-      .concat(currentLevel.thisLevel);
+    const combinedLength =
+      currentLevel.anyLevel.length + currentLevel.thisLevel.length;
     let added = false;
     // see if node matches
-    for (let i = 0; i < combined.length; i++) {
-      const exp = combined[i];
+    for (let i = 0; i < combinedLength; i++) {
+      const exp =
+        i < currentLevel.anyLevel.length
+          ? currentLevel.anyLevel[i]
+          : currentLevel.thisLevel[i - currentLevel.anyLevel.length];
       if (
         (!exp[0].id || vNode.shadowId === currentLevel.parentShadowId) &&
         matchesExpression(vNode, exp[0])