--- a/lib/core/utils/query-selector-all-filter.js
+++ b/lib/core/utils/query-selector-all-filter.js
@@ -34,7 +34,7 @@ function matchExpressions(domTree, expressions, filter) {
   let currentLevel = createLocalVariables(
     vNodes,
     expressions,
-    [],
+    null,
     domTree[0].shadowId,
     recycledLocalVariables.pop()
   );
@@ -42,17 +42,18 @@ function matchExpressions(domTree, expressions, filter) {
 
   while (currentLevel.vNodesIndex < currentLevel.vNodes.length) {
     const vNode = currentLevel.vNodes[currentLevel.vNodesIndex++];
-    const childOnly = []; // we will add hierarchical '>' selectors here
-    const childAny = [];
+    let childOnly = null; // we will add hierarchical '>' selectors here
+    let childAny = null;
     const combinedLength =
-      currentLevel.anyLevel.length + currentLevel.thisLevel.length;
+      (currentLevel.anyLevel?.length || 0) +
+      (currentLevel.thisLevel?.length || 0);
     let added = false;
     // see if node matches
     for (let i = 0; i < combinedLength; i++) {
       const exp =
-        i < currentLevel.anyLevel.length
+        i < (currentLevel.anyLevel?.length || 0)
           ? currentLevel.anyLevel[i]
-          : currentLevel.thisLevel[i - currentLevel.anyLevel.length];
+          : currentLevel.thisLevel[i - (currentLevel.anyLevel?.length || 0)];
       if (
         (!exp[0].id || vNode.shadowId === currentLevel.parentShadowId) &&
         matchesExpression(vNode, exp[0])
@@ -72,18 +73,18 @@ function matchExpressions(domTree, expressions, filter) {
           }
           if (rest[0].combinator === '>') {
             // add the rest to the childOnly array
-            childOnly.push(rest);
+            (childOnly = childOnly || []).push(rest);
           } else {
             // add the rest to the childAny array
-            childAny.push(rest);
+            (childAny = childAny || []).push(rest);
           }
         }
       }
       if (
         (!exp[0].id || vNode.shadowId === currentLevel.parentShadowId) &&
-        currentLevel.anyLevel.includes(exp)
+        currentLevel.anyLevel?.includes(exp)
       ) {
-        childAny.push(exp);
+        (childAny = childAny || []).push(exp);
       }
     }