@sorin-davidoi 2021-06-29
Opened it just to see if the CI is happy (https://github.com/dequelabs/axe-core/pull/3018#issuecomment-870064753), closing as it still doesn't run.